requirements.txt
pydantic>=2.6
pyyaml
fastapi
uvicorn[standard]
openai>=1.13
